+++ tests/Silverback.Integration.Tests/Messaging/Outbound/Enrichers/GenericOutboundHeadersEnricherFixture.cs
@@ -12,47 +12,29 @@ using Xunit;
 
 namespace Silverback.Tests.Integration.Messaging.Outbound.Enrichers;
 
-public class GenericOutboundHeadersEnricherTests
+public class GenericOutboundHeadersEnricherFixture
 {
     [Fact]
-    public void Enrich_StaticValues_HeaderAdded()
+    public void Enrich_ShouldAddHeaderViaValueProvider()
     {
         OutboundEnvelope<TestEventOne> envelope = new(
-            new TestEventOne(),
+            new TestEventOne { Content = "content" },
             null,
             TestProducerEndpoint.GetDefault(),
             Substitute.For<IProducer>());
 
-        GenericOutboundHeadersEnricher enricher = new("x-test", "value");
+        GenericOutboundHeadersEnricher<TestEventOne> enricher = new(
+            "x-test",
+            envelopeToEnrich => envelopeToEnrich.Message?.Content);
 
         enricher.Enrich(envelope);
 
         envelope.Headers.Should().HaveCount(1);
-        envelope.Headers.Should().BeEquivalentTo(new[] { new MessageHeader("x-test", "value") });
+        envelope.Headers.Should().BeEquivalentTo(new[] { new MessageHeader("x-test", "content") });
     }
 
     [Fact]
-    public void Enrich_StaticValues_HeaderReplaced()
-    {
-        OutboundEnvelope<TestEventOne> envelope = new(
-            new TestEventOne(),
-            new MessageHeaderCollection
-            {
-                { "x-test", "old-value" }
-            },
-            TestProducerEndpoint.GetDefault(),
-            Substitute.For<IProducer>());
-
-        GenericOutboundHeadersEnricher enricher = new("x-test", "value");
-
-        enricher.Enrich(envelope);
-
-        envelope.Headers.Should().HaveCount(1);
-        envelope.Headers.Should().BeEquivalentTo(new[] { new MessageHeader("x-test", "value") });
-    }
-
-    [Fact]
-    public void Enrich_SpecificMessageType_HeaderAddedToMessagesOfMatchingType()
+    public void Enrich_ShouldAddHeaderToMatchingMessageType()
     {
         OutboundEnvelope<TestEventOne> envelopeEventOne = new(
             new TestEventOne(),
@@ -75,7 +57,7 @@ public class GenericOutboundHeadersEnricherTests
     }
 
     [Fact]
-    public void Enrich_SpecificBaseMessageType_HeaderAddedToMessagesOfMatchingType()
+    public void Enrich_ShouldAddHeaderToMatchingBaseMessageType()
     {
         OutboundEnvelope<TestEventOne> envelopeEventOne = new(
             new TestEventOne(),
@@ -102,24 +84,5 @@ public class GenericOutboundHeadersEnricherTests
         envelopeEventOne.Headers.Should().HaveCount(1);
         envelopeEventTwo.Headers.Should().HaveCount(1);
         envelopeBinaryMessage.Headers.Should().BeEmpty();
-    }
-
-    [Fact]
-    public void Enrich_ValueProvider_HeaderAdded()
-    {
-        OutboundEnvelope<TestEventOne> envelope = new(
-            new TestEventOne { Content = "content" },
-            null,
-            TestProducerEndpoint.GetDefault(),
-            Substitute.For<IProducer>());
-
-        GenericOutboundHeadersEnricher<TestEventOne> enricher = new(
-            "x-test",
-            envelopeToEnrich => envelopeToEnrich.Message?.Content);
-
-        enricher.Enrich(envelope);
-
-        envelope.Headers.Should().HaveCount(1);
-        envelope.Headers.Should().BeEquivalentTo(new[] { new MessageHeader("x-test", "content") });
     }
 }